requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
boto3==1.42.17
botocore==1.42.17
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Events==0.5
granian==2.7.1
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
packaging==25.0
platformdirs==4.9.1
protobuf==6.33.5
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
redis==7.1.1
reflex==0.8.26
reflex-hosting-cli==0.1.61
requests==2.32.5
requests-aws4auth==1.3.1
rich==14.3.2
s3transfer==0.16.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.46
sqlmodel==0.0.33
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
watchfiles==1.1.1
wrapt==2.1.1
wsproto==1.3.2
